--- api/pvt/requirements.txt
@@ -1,4 +1,0 @@
-Flask==3.0.2
-Flask-Cors==4.0.0
-pymongo==4.6.3
-python-dotenv==1.0.1
