+++ api/requirements.txt
@@ -1,0 +1,8 @@
+Flask==3.0.2
+Flask-Cors==4.0.0
+pymongo==4.6.3
+python-dotenv==1.0.1
+py-vapid==1.9.0
+pywebpush==2.0.0
+python-dateutil==2.9.0.post0
+PyJWT==2.10.1
